ARM: Core support for booting SMP kernel on UP hardware
For SMP on UP let's forget about the kernel text being read-only.
Most of the SMP dependencies are in assembly - and we can make lists
of instructions and their modified versions that would be necessary
to boot a SMP kernel on UP.

So something like this will do (though note that not everywhere has been
fixed up - such as the page table flags.

If we don't want the SMP-on-UP support for SMP kernels (it's not actually
all that big - around 512 bytes) then we can discard the .smpalt.init
section and the __fixup_smp code.

Patch updated to merge in "Fix smp_on_up to not trash r1 and r2" patch.