staging: tidspbridge: make wake_dsp to handle PM code
Instead of being a wrapper around sm_interrupt_dsp, make wake_dsp
to be the one actually waking up the dsp and handling clock
configurations. No impact for sm_interrupt_dsp since it now calls
wake_dsp and that function decides what to do based on internal
states.

This should help in encapsulating code capable of sleep/wake the
dsp into a single place.

While at it, change 'if' blocks to a 'switch'; and use
omap_mbox_msg_send to send the wake up command to the dsp.

Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>