omap: Set separate timer init functions to avoid cpu_is_omap tests
This is needed for the following patches to initialize non-sys_timer
hardware later on. Again, there's no need to do cpu_is_omap calls
as we only care about the major omap generation. We can do this with
the .timer entry alone.

Note that for now we just set stubs for the various sys_timer
entries that will get populated in a later patch. Also the following
patches will remove the omap_dm_timer_init calls and make it an
arch_initcall.

Signed-off-by: Tony Lindgren <tony@atomide.com>
Reviewed-by: Kevin Hilman <khilman@ti.com>