powerpc: use irq_alloc_desc() to manage irq allocations
This patch drops the architecture specific code for managing irq
assignments and uses core code instead.

*RFC*

This patch is *not* ready for merging.  The locking is messed up where
the irq_map is still used to find out if a mapping is already established,
but irq_alloc_desc() is now called before the irq_map is updated which
creates a race condition.  I've not completely sorted out how to solve
this yet.  The solution may very well be to convert the irq_big_lock
to a mutex (but I'm not sure if anything that sleeps calls the affected
code), and hold the mutex over the entire operation.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>