powerpc: Make struct irq_host semi-private by moving into irqhost.h
Very few files actually need direct access to struct irq_host members.
This patch moves the irq_host definition into another file so that it
isn't brought in by default, and to prepare for the addition of
struct of_irq_domain, which will factor some of the irq_host behaviour
out into common code.  This needs to be done because of_irq_domain
will be embedded inside struct irq_host, and to do that it needs to be
guaranteed that struct of_irq_domain gets fully defined first.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>